refactor: common에 enums 패키지 만들어서 용돈일기 타입을 이동 시킴
그 외엔 임포트문 삭제
