+++ common/src/main/java/com/paykidscompose/common/model/allowance/CategoryModel.kt
@@ -1,6 +1,6 @@
 package com.paykidscompose.common.model.allowance
 
-import com.paykidscompose.common.model.AllowanceType
+import com.paykidscompose.common.enums.AllowanceType
 import com.paykidscompose.common.model.Model
 
 data class CategoryModel(
