+++ data/src/main/java/com/paykidscompose/data/mapper/allowance/AllowanceChartAmountMapper.kt
@@ -1,12 +1,13 @@
 package com.paykidscompose.data.mapper.allowance
 
+import com.paykidscompose.common.enums.AllowanceType
 import com.paykidscompose.common.mapper.ModelMapper
-import com.paykidscompose.common.model.AllowanceType
 import com.paykidscompose.common.model.allowance.AllowanceChartAmountModel
 import com.paykidscompose.data.model.allowance.AllowanceChartAmountDTO
 import java.time.LocalDate
 
-object AllowanceChartAmountMapper: ModelMapper<AllowanceChartAmountModel, AllowanceChartAmountDTO> {
+object AllowanceChartAmountMapper :
+    ModelMapper<AllowanceChartAmountModel, AllowanceChartAmountDTO> {
     override fun mapToModel(layerModel: AllowanceChartAmountDTO): AllowanceChartAmountModel {
         return AllowanceChartAmountModel(
             date = LocalDate.parse(layerModel.date),
