+++ data/src/main/java/com/paykidscompose/data/mapper/allowance/AllowanceChartCategoryMapper.kt
@@ -1,11 +1,12 @@
 package com.paykidscompose.data.mapper.allowance
 
+import com.paykidscompose.common.enums.AllowanceType
 import com.paykidscompose.common.mapper.ModelMapper
-import com.paykidscompose.common.model.AllowanceType
 import com.paykidscompose.common.model.allowance.AllowanceChartCategoryModel
 import com.paykidscompose.data.model.allowance.AllowanceChartCategoryDTO
 
-object AllowanceChartCategoryMapper: ModelMapper<AllowanceChartCategoryModel, AllowanceChartCategoryDTO> {
+object AllowanceChartCategoryMapper :
+    ModelMapper<AllowanceChartCategoryModel, AllowanceChartCategoryDTO> {
     override fun mapToModel(layerModel: AllowanceChartCategoryDTO): AllowanceChartCategoryModel {
         return AllowanceChartCategoryModel(
             type = AllowanceType.valueOf(layerModel.allowanceType),
