+++ data/src/main/java/com/paykidscompose/data/repositories/ExpenseCategoryRepositoryImpl.kt
@@ -4,7 +4,6 @@ import com.paykidscompose.common.model.allowance.CategoryModel
 import com.paykidscompose.common.repositories.ExpenseCategoryRepository
 import com.paykidscompose.common.result.DataResourceResult
 import com.paykidscompose.data.mapper.allowance.CategoryMapper
-import com.paykidscompose.data.model.allowance.CategoryDTO
 import com.paykidscompose.data.network.NetworkModule
 import com.paykidscompose.data.network.service.ExpenseCategoryApiService
 import kotlinx.coroutines.flow.Flow
@@ -12,14 +11,14 @@ import kotlinx.coroutines.flow.flow
 
 class ExpenseCategoryRepositoryImpl(
     private val categoryApiService: ExpenseCategoryApiService = NetworkModule.provideExpenseCategoryApiService()
-): ExpenseCategoryRepository {
+) : ExpenseCategoryRepository {
 
     override suspend fun deleteExpenseCategory(category: String): DataResourceResult<Boolean> {
         return runCatching {
             categoryApiService.deleteExpenseCategory(category)
         }.fold(
-            onSuccess = { DataResourceResult.Success(it.data)},
-            onFailure = { DataResourceResult.Failure(it)}
+            onSuccess = { DataResourceResult.Success(it.data) },
+            onFailure = { DataResourceResult.Failure(it) }
         )
     }
 
@@ -27,8 +26,8 @@ class ExpenseCategoryRepositoryImpl(
         return runCatching {
             categoryApiService.saveExpenseCategory(category)
         }.fold(
-            onSuccess = { DataResourceResult.Success(it.data)},
-            onFailure = { DataResourceResult.Failure(it)}
+            onSuccess = { DataResourceResult.Success(it.data) },
+            onFailure = { DataResourceResult.Failure(it) }
         )
     }
 
@@ -41,8 +40,9 @@ class ExpenseCategoryRepositoryImpl(
                 val models = it.data.map { dto ->
                     CategoryMapper.mapToModel(dto)
                 }
-                emit(DataResourceResult.Success(models))},
-            onFailure = { emit(DataResourceResult.Failure(it))}
+                emit(DataResourceResult.Success(models))
+            },
+            onFailure = { emit(DataResourceResult.Failure(it)) }
         )
     }
 }
